move run_coros_threadsafe to its own module and share it with llm map
